% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribo2ampvis.R
\name{ribo2ampvis}
\alias{ribo2ampvis}
\title{Convert Ribotagger data to ampvis2 data input file}
\usage{
ribo2ampvis(file)
}
\arguments{
\item{file}{The RIbotagger output file you want to convert}
}
\description{
Convert Ribotagger data to ampvis2 data input file
}
